requirements.txt
Flask
google-cloud-logging